face-detector/requirements.txt
Flask==3.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy<2
python-dotenv==1.0.0
mtcnn==1.0.0
tensorflow==2.20.0
requests==2.31.0
